style: keep the `L` on whole numbers that meet an array
The two spellings are not symmetric: an R integer widens into whatever
category it meets, while a plain `1` is an R double that pulls an integer
array into the float category (`x_i32 - 1` is `f32`). So a whole number
keeps its `L` against a float array too.

A literal that meets nothing is the exception -- it settles on the default
of its own category, which is why `nv_rnorm(mean = 0, sd = 1)` and the
other real-valued parameters, bounds, thresholds and coefficients stay
plain.

+++ R/primitives.R
@@ -454,7 +454,7 @@ prim_concatenate <- new_primitive(
       axis_const <- stablehlo::r_to_constant(
         as.integer(axis - 1L),
         dtype = "i64",
-        shape = integer(0)
+        shape = integer(0L)
       )
       out <- rlang::exec(stablehlo::infer_types_concatenate, !!!vts, dimension = axis_const)[[1L]]
       out <- vt2at(out)
@@ -832,7 +832,7 @@ infer_cum <- function(x, axis) {
   if (rank == 0L) {
     cli_abort("{.arg x} must have at least one axis to accumulate along, but it is a scalar.")
   }
-  if (!checkmate::test_integerish(axis, lower = 1, upper = rank, len = 1L)) {
+  if (!checkmate::test_integerish(axis, lower = 1L, upper = rank, len = 1L)) {
     cli_abort("{.arg axis} must be a single integer in 1:{rank}, but is {.val {axis}}")
   }
   list(AbstractArray(
@@ -851,7 +851,7 @@ infer_cum_extreme <- function(x, axis) {
   if (rank == 0L) {
     cli_abort("{.arg x} must have at least one axis to accumulate along, but it is a scalar.")
   }
-  if (!checkmate::test_integerish(axis, lower = 1, upper = rank, len = 1L)) {
+  if (!checkmate::test_integerish(axis, lower = 1L, upper = rank, len = 1L)) {
     cli_abort("{.arg axis} must be a single integer in 1:{rank}, but is {.val {axis}}")
   }
   list(
@@ -2253,7 +2253,7 @@ prim_iota <- new_primitive(
       iota_axis_const <- stablehlo::r_to_constant(
         as.integer(axis - 1L),
         dtype = "i64",
-        shape = integer(0)
+        shape = integer(0L)
       )
       # Just for the checks
       stablehlo::infer_types_iota(iota_dimension = iota_axis_const, dtype = dtype, shape = shape)[[1L]]
